requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==9.5.0
python-dotenv==1.0.0
drf-yasg  # Swagger documentation